requirements.txt
PyYAML==6.0
types-PyYAML==6.0.12.20240808
psutil==6.0.0
types-psutil==6.0.0.20240901
pynput==1.7.7
types-pynput==1.7.7.20240813
pyinstaller==6.10.0